refactor(workflows): move the workflow seam out of index.ts
The workflow feature landed ~1000 lines in index.ts, taking it to 4374 —
past the point where reading it in full before a wide-ranging change is
affordable. The engine under src/workflow/ was already well factored; the
seam file was not.

Four moves, all behaviour-preserving:

- src/ui/workflow-menu.ts takes showWorkflowDialog, openWorkflowFromFleet
  and showWorkflowsMenu. /agents has four submenus and three already
  followed schedule-menu.ts; the workflow one was the exception. Deps
  arrive as WorkflowMenuDeps rather than through the factory closure.
- src/workflow/collisions.ts takes FOREIGN_WORKFLOW_TOOL_NAMES and a pure
  decideWorkflowCollision(). The policy — what counts as a conflict, what
  a pin changes, whether anything is left to withdraw — is now separable
  from the host half that reads getAllTools and calls setActiveTools.
- src/workflow/entry.ts takes WORKFLOW_ENTRY_TYPE and workflowEntryData,
  so the shape a session file stores no longer imports the renderer that
  draws it; renderWorkflowEntryCard moves beside renderWorkflowCard.
- The pure resolvers go to the modules that own their topic:
  resolveWorkflowScript to saved.ts beside resolveWorkflowSource,
  resolveResumeTarget and formatWorkflowNotification to task.ts,
  workflowCallName to meta.ts.

escapeXml gets its own module rather than moving into task.ts: both
notification builders need it and they sit on opposite sides of an import
edge, so the workflow one cannot reach back into index.ts for it.

fleetWorkflows deliberately stays. Moving it would make the engine import
FleetWorkflow from ui/fleet-list.js, inverting the ignorance the fleet
list is documented to keep. It is eleven lines and it is the seam.
